+++ Packages/Newtonsoft.Json-for-Unity.Converters/UnityConverters/Math/Vector3IntConverter.cs
@@ -27,22 +27,34 @@ using UnityEngine;
 namespace Newtonsoft.Json.UnityConverters.Math
 {
     /// <summary>
-    /// Custom Newtonsoft.Json converter <see cref="JsonConverter"/> for the Unity Vector3Int type <see cref="Vector3Int"/>.
+    /// Custom Newtonsoft.Json converter <see cref="JsonConverter"/> for the Unity integer Vector3 type <see cref="Vector3Int"/>.
     /// </summary>
-    public class Vector3IntConverter : PartialIntConverter<Vector3Int>
+    public class Vector3IntConverter : PartialConverter<Vector3Int>
     {
-        public Vector3IntConverter() : base(Vector3Converter._memberNames)
+        protected override void ReadValue(ref Vector3Int value, string name, JsonReader reader, JsonSerializer serializer)
         {
+            switch (name)
+            {
+                case nameof(value.x):
+                    value.x = reader.ReadAsInt32() ?? 0;
+                    break;
+                case nameof(value.y):
+                    value.y = reader.ReadAsInt32() ?? 0;
+                    break;
+                case nameof(value.z):
+                    value.z = reader.ReadAsInt32() ?? 0;
+                    break;
+            }
         }
 
-        protected override Vector3Int CreateInstanceFromValues(ValuesArray<int> values)
+        protected override void WriteJsonProperties(JsonWriter writer, Vector3Int value, JsonSerializer serializer)
         {
-            return new Vector3Int(values[0], values[1], values[2]);
-        }
-
-        protected override int[] ReadInstanceValues(Vector3Int instance)
-        {
-            return new int[] { instance.x, instance.y, instance.z };
+            writer.WritePropertyName(nameof(value.x));
+            writer.WriteValue(value.x);
+            writer.WritePropertyName(nameof(value.y));
+            writer.WriteValue(value.y);
+            writer.WritePropertyName(nameof(value.z));
+            writer.WriteValue(value.z);
         }
     }
 }