+++ Packages/Newtonsoft.Json-for-Unity.Converters/UnityConverters/PartialConverter.cs
@@ -23,10 +23,7 @@
 #endregion
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Reflection;
 using Newtonsoft.Json.UnityConverters.Helpers;
 
 namespace Newtonsoft.Json.UnityConverters
@@ -34,78 +31,12 @@ namespace Newtonsoft.Json.UnityConverters
     /// <summary>
     /// Custom base <c>Newtonsoft.Json.JsonConverter</c> to filter serialized properties.
     /// </summary>
-    /// 
-    /// <remarks>
-    /// Useful for Unity or 3rd party classes, since we can't insert any <c>Newtonsoft.Json.JsonIgnoreAttribute</c>.
-    /// By the way, this works by reflection to access properties.
-    /// Please make sure your property not to be stripped by Unity.
-    /// </remarks>
-    /// 
-    /// <example>
-    /// It's very easy to make a custom converter, just inherit and override <c>GetPropertyNames()</c> as the filter:
-    /// </example>
-    /// 
-    /// <code>
-    /// public class SomeConverter : PartialConverter<SomeClass>{
-    /// 	protected override string[] GetPropertyNames(){
-    /// 		return new []{"someField", "someProperty", "etc"};
-    /// 	}
-    /// }
-    /// </code>
-    /// 
-    public abstract class PartialConverter<T, TInner> : JsonConverter
+    public abstract class PartialConverter<T> : JsonConverter
+        where T : new()
     {
-        private readonly Dictionary<string, int> _namesIndices;
-        private readonly string[] _namesArray;
+        protected abstract void ReadValue(ref T value, string name, JsonReader reader, JsonSerializer serializer);
 
-        /// <summary>
-        /// Initializes this converter with the set of fields/properties to
-        /// read/write to the object.
-        /// </summary>
-        /// <param name="propertyNames">The list of values.</param>
-        protected PartialConverter(string[] propertyNames)
-        {
-            _namesArray = propertyNames.ToArray(); // Intentionally make a copy of the array
-
-            _namesIndices = new Dictionary<string, int>(_namesArray.Length);
-
-            for (int i = 0; i < _namesArray.Length; i++)
-            {
-                _namesIndices[_namesArray[i]] = i;
-            }
-        }
-
-        /// <summary>
-        /// Create the instance with the given values.
-        /// </summary>
-        /// <param name="values">The values read from the object. Known to have the same size as number of elements fed through the constructor of this PartialConverter.</param>
-        /// <returns>The instance.</returns>
-        protected abstract T CreateInstanceFromValues(ValuesArray<TInner> values);
-
-        /// <summary>
-        /// Read the values off from the given instance.
-        /// The returned list must have the same number of elements as elements fed through the constructor.
-        /// </summary>
-        /// <param name="instance">The instance to read the values off.</param>
-        /// <returns>The values.</returns>
-        protected abstract TInner[] ReadInstanceValues(T instance);
-
-        /// <summary>
-        /// Writes a value directly to the JSON writer. Meant to implement the appropriate WriteValue
-        /// <see cref="JsonWriter.WriteValue(object)"/>
-        /// for the generic type.
-        /// </summary>
-        /// <param name="writer">The JSON writer</param>
-        /// <param name="value">The value to write</param>
-        protected abstract void WriteValue(JsonWriter writer, TInner value, JsonSerializer serializer);
-
-        /// <summary>
-        /// Read a value directly from the JSON reader. Meant to implement the appropriate ReadAsX,
-        /// (ex: <see cref="JsonReader.ReadAsInt32()"/>)
-        /// for the generic type.
-        /// </summary>
-        /// <param name="reader">The JSON reader</param>
-        protected abstract TInner ReadValue(JsonReader reader, int index, JsonSerializer serializer);
+        protected abstract void WriteJsonProperties(JsonWriter writer, T value, JsonSerializer serializer);
 
         /// <summary>
         /// Determine if the object type is <typeparamref name="T"/>
@@ -135,21 +66,20 @@ namespace Newtonsoft.Json.UnityConverters
             [AllowNull] object existingValue,
             JsonSerializer serializer)
         {
-            bool isNullableStruct = objectType.IsGenericType
-                && objectType.GetGenericTypeDefinition() == typeof(Nullable<>);
+            if (reader.TokenType == JsonToken.Null)
+            {
+                bool isNullableStruct = objectType.IsGenericType
+                    && objectType.GetGenericTypeDefinition() == typeof(Nullable<>);
 
-            return InternalReadJson(reader, serializer, isNullableStruct);
+                return isNullableStruct ? null : (object)default(T);
+            }
+
+            return InternalReadJson(reader, serializer);
         }
 
         [return: MaybeNull]
-        private object InternalReadJson(JsonReader reader, JsonSerializer serializer, bool isNullableStruct)
+        private T InternalReadJson(JsonReader reader, JsonSerializer serializer)
         {
-
-            if (reader.TokenType == JsonToken.Null)
-            {
-                return CreateValueForNull(isNullableStruct);
-            }
-
             if (reader.TokenType != JsonToken.StartObject)
             {
                 throw reader.CreateSerializationException($"Failed to read type '{typeof(T).Name}'. Expected object start, got '{reader.TokenType}' <{reader.Value}>");
@@ -157,21 +87,20 @@ namespace Newtonsoft.Json.UnityConverters
 
             reader.Read();
 
-            var values = new ValuesArray<TInner>(_namesArray.Length);
-            int previousIndex = -1;
-
+            var value = new T();
+            string previousName = null;
+            
             while (reader.TokenType == JsonToken.PropertyName)
             {
-                if (reader.Value is string name
-                    && _namesIndices.TryGetValue(name, out int index))
+                if (reader.Value is string name)
                 {
-                    if (index == previousIndex)
+                    if (name == previousName)
                     {
                         throw reader.CreateSerializationException($"Failed to read type '{typeof(T).Name}'. Possible loop when reading property '{name}'");
                     }
-
-                    previousIndex = index;
-                    values[index] = ReadValue(reader, index, serializer);
+                    
+                    previousName = name;
+                    ReadValue(ref value, name, reader, serializer);
                 }
                 else
                 {
@@ -181,21 +110,7 @@ namespace Newtonsoft.Json.UnityConverters
                 reader.Read();
             }
 
-            return CreateInstanceFromValues(values);
-        }
-
-        [return: MaybeNull]
-        private object CreateValueForNull(bool isNullableStruct)
-        {
-            if (isNullableStruct)
-            {
-                return null;
-            }
-            else
-            {
-                var values = new ValuesArray<TInner>(_namesArray.Length);
-                return CreateInstanceFromValues(values);
-            }
+            return value;
         }
 
         /// <summary>
@@ -215,45 +130,9 @@ namespace Newtonsoft.Json.UnityConverters
             writer.WriteStartObject();
 
             var typed = (T)value;
-
-            TInner[] values = ReadInstanceValues(typed);
-
-            if (values?.Length != _namesArray.Length)
-            {
-                throw writer.CreateWriterException(string.Format("Expected {0}() to return {1} values, matching [{2}]. Got {3}",
-                    nameof(ReadInstanceValues),
-                    _namesArray.Length,
-                    string.Join(", ", _namesArray),
-                    values?.Length.ToString() ?? "null")
-                );
-            }
-
-            for (int i = 0; i < _namesArray.Length; i++)
-            {
-                string name = _namesArray[i];
-                writer.WritePropertyName(name);
-                WriteValue(writer, values[i], serializer);
-            }
+            WriteJsonProperties(writer, typed, serializer);
 
             writer.WriteEndObject();
-        }
-
-        /// <summary>
-        /// Gets the non-public instance field info <see cref="FieldInfo"/> for the converted type
-        /// <typeparamref name="T"/>.
-        /// If not found then will throw a missing member exception <see cref="MissingMemberException"/>.
-        /// </summary>
-        /// <remarks>
-        /// If used in static initialization (ex: inside static constructor,
-        /// static field, or static property backing field initialization)
-        /// and the field does not exist it would invalidate the type for
-        /// the entirety of the programs lifetime.
-        /// </remarks>
-        /// <param name="name">Name of the non-public instance field.</param>
-        protected internal static FieldInfo GetFieldInfoOrThrow(string name)
-        {
-            return typeof(T).GetField(name, BindingFlags.NonPublic | BindingFlags.Instance)
-                ?? throw new MissingMemberException(typeof(T).FullName, name);
         }
     }
 }